roles/xmpp: Downgrade TLS ciphers
Switch back to Prosody 0.9.7 defaults, minus non-forward-secret
ciphers.

We aren't using compression or CBC-mode ciphers, so TLSv1 and 1.1
ciphers should be safe. Also, we allow ciphers using SHA1 MAC for
now, as it seems to be the prevailing wisdom to enable support older
clients. Anyways, we enforce the server's order of ciphers so real
clients will be ok.

Signed-off-by: Alan Orth <alan.orth@gmail.com>